Always use size_t for MTU

+++ src/iface.h
@@ -1,3 +1,3 @@
 #pragma once
 
-int iface_set_mtu (const char *, int);
+int iface_set_mtu (const char *, size_t);
